Extracts format message
